external/asf4: Upgrade to latest version
A new project was created in Atmel Start (start.atmel.com) and all
drivers we are using was added and our clock configuration was setup.

After that the content of the asf4-drivers directory was removed and the
new project was unpacked in its place.

Finally the qtouch and sdcard drivers were reverted as they have been
modified by us.

NOTE: Some code was present in the project that wasn't used, like an
USART and an AES driver.

NOTE: The ASF4 drivers doesn't take `const` pointers when writing
buffers, so we need to cast from `const` to non-`const`. This is cought
by GCC, so the check was removed from clang-tidy.

NOTE: Moved some of our additions to the `src` directory. Like
sd_mmc_ext.c