api: add Unlock call w/ support for entering passphrase on the host
Previously, the host would unlock before the noise channel was
established using `OP_UNLOCK`, using a raw unencrypted call that is using
protobufs.

To support entering a passphrase on the host, we add a `UnlockRequest`
protobuf message. This needs to happen after the noise channel was
established because:
- we want the passphrase to be encrypted in flight
- we want to make use of protobuf and `next_request()` to query the
  host

`UnlockRequest` has a flag `supports_host_passphrase`, so
wallets (especially third party apps) that do not support this yet
still work seamlessly, as without support the passphrase is simply
entered on the device.

If the host app supports it, then the device will ask the user if they
want to enter on the device or on the host.

The query to the host uses `UnlockRequestHostInfoResponse`. It
contains the `type` enum for future compatbilitiy in case we want to
add more unlock options in the future (e.g. fingerprints).

The `workflow::unlock` calls now get a callback to pick the method of
entering the passphrase, so the old `OP_UNLOCK` can use the previous
way, while `UnlockRequest` provides a callback to ask the user where
to enter and proceeds accordingly.