rust: remove SafeInputString
This was a custom C string wrapper with auto-zeroing, but it is much
easier to simply use zeroize and CStr (which were both introduced to
the codebase after SafeInputString). This also allows us to call bip39
unlock more easily with a Rust string in the future (see testing.rs
for example).