factory-setup: Remove C implementation of RTT
The debug builds use the rust implementation of RTT. The two
implementations cannot coexist. To be able to build debug builds of the
factory-setup the C imeplementation is therefore removed.

The factory-setup image is changed to use the second channel for API
calls, since the first is used for debug console.