make blocking workflows before the mainloop async
The last two remaining non-async workflows are now made async. This
will help with BB02+ Bluetooth chip integration, whose firmware we
want to stream in an async fashion at boot while the orientation
screen and idle workflows are running, to not cause a noticable delay
before or after these workflows.