memory: return unique default device name for BB02+
To disambiguate multiple BitBoxes in the Bluetooth devices list, we
use a unique name of the form "BitBox ABCD" (four random uppercase
letters) until a device name is set by the user.

The random letters are not picked perfectly uniformely, but this is
not security critical - it's merely to create a unique name to be able
to disambiguate.

This change only applies to BB02+, and this random device name is
displayed on the lockscreen visible before pairing.