memory: validate name when setting/getting
While the BitBox02 API memory_setup() and SetDeviceName API checks
should make sure that `memory_get_device_name()` always returns a
valid ASCII-encoded string, we add explicit checks to make sure it is
so.

The reason for this is that the name is sent over to the Bluetooth
firmware in the BB02+, and an invalid string there could cause a
crash.