ui: use a trait for workflows
This allows us to plug in a local trait impl for unit tests, allowin
us to remove the ugly unsafe static muts, RefCells and UI_COUNTERS in
the testing code and makes UI unit tests cleaner.

Another benefit of this is that the API code would not hardcode the UI
workflows, so future BitBox products with different UI can use the
same API code but plug its own UI workflows.

We do this by using a `Workflows` trait param in all functions that
need workflows, and use an implementation suitable for unit tests.

One instance of `RealWorkflows` is created that can be added as an arg
in any API function.

This commit is not a complete migration, we can migrate file by file
or function by function. The end goal is to remove all direct calls to
`workflow::*`.