api/ethereum: convert tests to Workflows trait, fix cancellation bug
The bug was that "Transaction canceled" was only shown then the
recipient or total fees were cancelled. If one cancelled the chain
confirmation (e.g. "Sign transaction on Ethereum"), the "Transaction
canceled" status was not shown.

It is fixed by doing it the same way as in bitcoin/signtx.rs: wrapping
the process function.