keystore: port secp256k1_schnorr_bip86_pubkey to Rust
Since the inclusion of the miniscript feature and rust-miniscript dep,
we depend on rust-bitcoin and rust-secp256k1, so we can make use of it
instead of manually wrapping secp256k1. This does not change the
binary size.

The deleted C test was already previously replicated in
`test_schnorr_bip86_pubkey` in bip32.rs.