port cipher to Rust
Until the relevant keystore.c functions are ported too, we keep the
cipher.c/cipher.h interface but have it call into Rust.

The test_cipher.c tests were ported and the test generator script
adatped to produce Rust output.

debug-assertions are enabled so `cargo test` can trigger the
`debug_assert!()` calls, which we don't want to be active in the
released firmwares.