factorysetup: fix OP_BLE_RESULT
output[0] is the command, output[1] is the the error according to
`error_code_t`. For sanity, the OP_BLE_RESULT call should just return
the ble_error_code_t result var, so we put it into
output[2]. output[1] will be 0, meaning fetching the BLE result
suceeded, even if that result itself might be a failure.