fonts: Make fonts closer to original
Fonts are generated with `ugui.h` as the path to `ugui.h`. We add the
full path to `ugui.h` as an include directory which let's us use fonts
with fewer manual modifications.

UG_FONT_DATA is originally defined as `const`. Not sure why this was
changed.

Bump the tool so it supports homebrew and use our constant for
`UG_FONT_DATA`.