ble: Wait with setting the product string to after orientation
usb_processing wasn't being setup until AFTER usb was enabled. This was
a problem beacuse the app is so fast that it manages to send the first
requests over ble before usb is setup.

Since "usb_processing" is unrelated to the usb driver, we initialize it
before we initialize the usb driver.