usb_processing: increase watchdog timeout once
The firmware typically expects keep-alive requests every 500ms from the
app. However keystore unlock takes about 5-6s so for that single call we
reset the watchdog counter to a lower value allowing approximately a
7s gap between messages from the app.

In practice this has worked anyway for the usb stack due to the order of
things in the main loop. I.e. check the next incoming message before
checking the timeout counter.