U2F: Fix on safari/macos
U2F is typically a stateless protocol, so between two transactions
(request+response) there is not supposed to be any lock held. It was
assumed that the "FIDO Client" (typically web browser) would keep
sending only U2F `register` or `authenticate` messages when it had started
doing so. But it is also legal to send for example a new U2FHID `init`
message in between. (the U2F reference test cases do this for example).

Before this commit we held a "lock" while the confirm screen was active
until the user had confirmed and didn't allow any other messages in
between. This breaks the U2F protocol as explained above.

After this commit we instead only hold the "lock" for the transaction
(request + response), which is the correct way according to the u2f
reference:

    The application channel that manages to get through the first
    initialization packet when the device is in idle state will keep the
    device locked for other channels until the last packet of the
    response message has been received. The device then returns to idle
    state, ready to perform another transaction for the same or a
    different channel. Between two transactions, no state is maintained
    in the device and a host application must assume that any other
    process may execute other transactions at any time.