workflow: default to digits in password entry when using Optiga
The Optiga securechip implementation limits password unlock attempts
to 10 at the secure chip config level, whereas in the ATECC, the
securechip-imposed limit is ~700k. For this reason, when using the
Optiga, users can safely use passwords consisting only of digits, so
we default to the digits keyboard in that case. Letters are still
available through the keyboard switcher.