factory-setup: Check if BLE chip is booted
If the BLE chip boots for the first time it will request various things
by itself, like the bond db. But if the chip already is booted (in case
factory setup is run twice) we need to request something to get a
message from it.

The main MCU has to respond immediately when it sees STX, loading the
firmware must be done before we reset the BLE MCU so that this MCU is
ready.