test: remove _test_secp256k1_schnorr_sign
It is not a unit test of our function, but the secp256k1
function, and was a sanity check during development.

This is already tested in the secp256k1 C and Rust libraries, and the
unit tests of `keystore_secp256k1_schnorr_sign()` and the transitive
tests in signtx.rs are sufficient to catch regressions.