bip32: use bitcoin::bip32 for xpub derivation
bip32 derivation involves hmac-sha512, and bitcoin::bip32 uses the
bitcoin-hashes crate for this, pulling in a second implementation of
sha512 (next to wally_sha512) and another hmac implementation (next to
wally's hmac and the hmac crate). This adds ~4kB of binary space, but
we should get most of it back when we remove libwally (specifically
wally-bip39, which also uses hmac-sha512).