keystore: remove _get_prv
It is now in `bitbox02_rust::keystore`. The remaining use of it in
keystore.c is to derive the private key at a keypath, which we can do
by exposing the Rust function we have for that to C. With this commit,
the last use of libwally for bip32 derivation is gone.