fix null ptr deref
`confirm_gesture_create` was called with `NULL` as parent, so
`ui_util_position_right_top(parent, confirm_gesture);` passed NULL,
and the function derefed NULL to compute the position.

Though undefined behavior technically, 0 is a valid pointer in the
BitBox02 (start of bootloader), so some bytes slightly offset from 0
were used for the positions.

None of that impacted the device behavior though as the positioning of
the component was not needed in the first place. The component has a
custom _render function that renders the confirm arrows in the right
location without looking at `component->position`, so we can just
remove the offendling line.

The parent param is removed as it is unused.