ui/trinary_input_string: directly fetch bip39 words
Previously the component was passed an array of string pointers, which
worked well with libwally's static BIP39 word list.

When we want to move to a Rust bip39 wordlist, we can't directly pass
Rust word string pointers (not null terminated).

This commit adapts the component to call a function to get a BIP39
word by index, which can be replaced with a Rust implementation later.