remove the option of restoring from 18 recovery words
We don't offer creating such mnemonics, and it seems it is not adopted
by the wider industry/community. Removing it leads to better UX.

These calls are the only ones, so we could actually drop the middle
option totally and rename the component to `binary_choice`, but we
keep it like this for a while in case we want to revert the change.