build: fix sanitizers
The recommended way to enable sanitizers nowadays is to add
`-fsanitize=*` as a compiler and linking flags.

Also fix some warnings and remove some targets from "all" to speed up
builds.

Bumping macosx version to 11 since that seems to be necessary for asan
to work.

Add CI check to verify that address santizier is turned on for unit
tests