confirm_gesture: fix unintentional keyboard press
Previously the confirm gesture component would get "out of active" as
soon as the top slider was released. If one released the top slider
before the top slider it would often trigger a keyboard press.

This commit creates an additional boolean that keeps the "active" state
in "active" until the drawing is reset to it's original state thus
removing the risk of unintentional keyboard presses.

The confirm gesture is also given a "higher priority" by being rendered
last.