avoid flicker when changing from lockscreen to waiting screen
The BitBox shows the lockscreen ("See the BitBoxApp") first, and when
the noise connection is initiazed, switches to the BitBox logo, which
stays there as the waiting/default screen until reconnected.

The noise connection is initiated by the host after unlock finishes,
so in between, "See the BitBoxApP" flashes for a split second.

This commit solves this by

1) not pushing the lockscreen as an extra component to be popped, but
folding it into the top/default component

2) updating that screen when the app initiates the connection if the
device is not initialized

3) if initialized, updating the screen on OP_UNLOCK (see comment in code)

This prevents flashing/flickering if the device is initialized. There
is still a small flicker when the device is uninitialized, as the
screen is updated before the pairing confirmation appears. This was
the case before this commit too.