hal: make SD functions async and yielding
An SD operation resume clock, mount, does the operation, unmounts,
stops clock. This is a bit slow. In API calls like backup creation,
this happens 10+ times in a row, which can cause communication
timeouts.

We make all SD functions async. Instead of making all of them properly
async, for now we make all of them yield to the executor afterwards,
so the communication processing gets a chance to run and not cause a
timeout. In the future we could then swap the sdcard functions with
proper async ones.