u2f: Do not turn off BLE when U2F is used over USB
iOS devices support u2f over USB, so no need to disable BLE when only
u2f is used.

iOS devices also send one packet of garbage when waking up, 64 null
bytes, which also is ignored, so that the bitbox still can connect to
the iOS device when it is awakened.