keystore: port stretch_retained_seed_encryption_key to Rust
- Replace function body by Rust impl, removing the original C function
and Rust wrapper
- Expose to C using extern "C" and use that in keystore.c

The function now returns bool instead of keystore_error_t for simpler
code. Among all the callsites, only one propagates keystore_error_t
instead of bool, and there we simply return a new error. We don't lose
much info, if it fails in the wild it is surely because of the
securechip and not the hashing.