memory: Check if hww reset was in progress during init
If the device is powered off after erasing but before initializing has
finalized it is bricked since a noise key is required.

Set a flag in chunk 7 that isn't being erased before reset and clear
that flag when reset is done. During boot if that flag still is set,
rerun the reset.