factorysetup: update comment about auto_enter/upside_down
The old comment was from a time where jlink was already locked when we
flashed this image, so it was required to be able to get into the
bootloader again. Now this is not the case anymore, but we still need
these two flags to be set properly like this regardless.