keystore: port bip39 unlock/retain to Rust
retain_bip39_seed encrypted the seed, copy_bip39_seed decrypted
it. This is ported to a helper struct `RetainedEncryptedBuffer` that
does the same. The helper struct is made because the other
seed (_retained_seed_encrypted) does the exact same operation, just
with a different purpose string when hashing, so we can reuse that
later when porting retain_seed/copy_seed.

_is_unlocked_bip39 is not ported, as it is redundant and equivaent with
RETAINED_BIP39_SEED being Some().