usb: use a static write buffer
Instead of relying on the caller to not modify the buffer a copy is made
to static memory. This enables the caller to use a stack allocated
buffer.

Because the USB stack uses DMA to move memory all buffers need to be
32 bit aligned. By coincidence, the buffer we currently use in the
"queue" were aligned. But when buffers are stack allocated there is a
higher chance that they sometimes are not 32 bit aligned