optiga: add initial optiga_test.c
The optiga sync ops functions are moved to a new file so they can be
faked in the unit test.

Some defines are moved from optiga.c to optiga.h to make them
available to the unit tests.

The unit test file itself fakes the optiga ops functions to a useful
degree.

The goal is to unit test the stretching algo based on fixtures, so we
can catch regressions.