gitignore: ignore .venv directories
The venv module supports creating lightweight "virtual environments",
each with their own independent set of Python packages usually
installed in a local .venv directory.

These packages, should not be commited an can be ignored.

For example in case a user sets up a python virtual environment for
the python scripts in py directory.