optiga: also include config update in unit tests
Some more defs are moved from optiga.c to optiga.h to make them
available to the tests.

By faking the metadata reading/writing of the two new slots, we can
exercise the code that updates and checks the config.

_reset_fakes is renamed to _setup_test and now includes
optiga_setup(), so we can check the config was updated.