py: fix fmt linter
Before a recent black update, black would, for each file it formatted,
apply the pyproject.toml related to it, e.g. py/pyproject.toml for all
files in py/, and not for files outside of py/. Since the black
update, if formatting files both inside and outside of py/, it would
ignore pyproject.toml, as it's not in the root.

We do a workaround by supplying configuring the linter to use
py/pyproject.toml for all files. Alternative would be to apply black
file-by-file, but that is much slower. Another alternative is to make
split it into two runs, but that was too much trouble. Moving
pyproject.toml from py/ to the root seemed wrong, as the root is not a
Py project.